Changed strategy for updating colors on scrolling--reduces sprite flicker.
Instead of using up sprites to mask wrong colors on the right side of the
screen, the wrong colors are now only 3-4 pixels on both the right and left
sides. This saves a little bit of sprite flickering, still with less wrong
colors visible than commercial games such as Super Mario Bros. 3.

I also fixed a bug/mis-feature in the scrolling subroutines: attribute table
updates were reusing the last value sent to the nmi nametable update routine,
despite having code meant to update it. This allowed nametable and attribute
table updates to happen on the same frame, but made them very dependant on
each other's timing. Scrolling from the right edge was written so that they
would always occur at the same time, and could not be made to stagger them
without changes elsewhere, which was required for the feature change above.
The updates could be made to run at the same time again if their parameters
were separated in ram, but staggering them should be better for performance.